Refactor dockerregistry: prevent a warning during startup
`REGISTRY_URL` is used by our middleware code to prefix image names in
the registry. It's unknown to registry's configuration parser which
tries to match it against a config option. Its failed attempt results in
a warning:

    level=warning msg="Ignoring unrecognized environment variable REGISTRY_URL"

This patch renames it to `DOCKER_REGISTRY_URL` and thus prevents the
warning.

Signed-off-by: Michal Minar <miminar@redhat.com>