Do not store layer links in the repository
There are several ways to lose links in the repository. User can delete
imagestreamtag or whole imagestream from etcd. So the `oc prune image`
will not help here (only hardprune can help when it learns to remove
them). At the same time, links are involved in access control. To avoid
possible harm, we must always use the information from the etcd to
resolve links. We have a global descriptor cache so we will not
constantly access to master api.

Signed-off-by: Gladkov Alexey <agladkov@redhat.com>