Use a static libtrust key to ensure consistent pullthrough manifests
The signing key in a schema1 manifest is effectively unused or verified,
and no client leverages them today. However, the use of a randomly generated
key ensures that subsequent pulls of a schema1 manifest via the v2 endpoint
will always return different body content, which defeats the ability of a
caller to verify the content they pulled.

Instead of generating every time, hardcode a specific key so that the body
content is stable and predictable. When schema1 support is removed the key
will be inert.