Checking if image name is valid, prune otherwise.
In some cases we have some images on the blob storage that contain names
considered to be invalid by the ImageStream API. Names containing / or
even % are not considered valid for Images.

As some customers are not able to prune their images due to this error
this patch adds a check and sets the image to be pruned if it contains
an invalid name(they should not be there anyways).