Disabled pprof and marking exec() as safe.
gosec is complaining about two issues. One is that we have pprof enabled
on our process and the other is about using variables when shelling out
commands. As far as I have checked there is no potential harm in using
exec the way we are using as command line arguments are properly escaped
before shelling out.

As for pprof, this patch disables it.