Implement retries when checking for blob presence.
This PR refers to https://bugzilla.redhat.com/show_bug.cgi?id=1745743.
We do a best effort here and retry when checking if a blob is present or
not. We need to do this as some of image registry storage backends have
eventual consistency and in some cases we may ended up checking for a
blob presence before it shows up.

This PR acts only on "manifest upload" as it tends to happen right
after the client uploaded blob layers. In the future we gonna need to
get rid of these retries, maybe by leveraging docker distribution redis
cache layer.