Wrap docker keyring in a auth.CredentialStore.
Added a package to wrap a docker keyring as returned by kubernetes
MakeDockerKeyring() into a docker distribution auth.CredentialStore.

This package is on the wild for a while now and has been copied from
openshift-apiserver into here.

Other than wrapping the keyring into a auth.CredentialStore this
package also makes multiple attempts when looking for auth credentials
for some registries.