Only warn if PublicDockerImageRepository is set
It is not an error for PublicDockerImageRepository to be set empty
to an image stream (it means the registry is not exposed by default).

It is also not an error for DockerImageRepository to be set if the
registry is currently starting up, but it is uncommon enough that
warning is a good idea.