pkg/dockerregistry/server: handle push by digest
to support manifest lists and oci image indexes, push by digest support
is required.

when a client pushes a manifest list, they will first push each
sub-manifest and its blobs, then finally push the manifest list.

a connection between an image and an image stream is represented via
image stream tags. an image which represents a sub-manifest of a
manifest list is however not directly connected to an image stream, and
it doesn't have an image stream tag.

because of this lack of image stream tag for images representing
sub-manifests of a manifest list, the registry needs to perform a
different action when faced with a push by digest. instead of creating
an image stream mapping between the image stream and the image, it will
simply create an image object for the sub-manifest.