OCPBUGS-38874: add ap-southeast-5 region
this pr adds the ap-southeast-5 region to the list of supported regions.
this change is made directly into the vendor directory and is intended
only for test purposes. if we get the ok that this works we will make
the changes and bumps necessary to land this in the right place.